+++ src/main/java/com/alwis/ecommerce/productcatalogapi/service/ProductServiceImpl.java
@@ -5,12 +5,15 @@ import com.alwis.ecommerce.productcatalogapi.model.Category;
 import com.alwis.ecommerce.productcatalogapi.model.Product;
 import com.alwis.ecommerce.productcatalogapi.repository.CategoryRepository;
 import com.alwis.ecommerce.productcatalogapi.repository.ProductRepository;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 
+@Slf4j
 @Service
 public class ProductServiceImpl implements ProductService{
 
@@ -22,6 +25,7 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public Product createProduct(Product product) {
+        log.info("Saving product: {}",product.getName());
         Long categoryId = product.getCategory().getId();
 
         Category category = categoryRepository.findById(categoryId).orElseThrow(() -> new RuntimeException("Category not found with ID: " + categoryId));
@@ -33,11 +37,14 @@ public class ProductServiceImpl implements ProductService{
 
     @Override
     public List<Product> getAllProducts() {
+        log.info("Fetching all products");
         return productRepository.findAll();
     }
 
     @Override
     public Product getProductById(Long id) throws ProductNotFoundException {
+
+        log.info("Fetching product with id:{}",id);
         Optional<Product> product = productRepository.findById(id);
 
         if(!product.isPresent()){
@@ -48,7 +55,14 @@ public class ProductServiceImpl implements ProductService{
     }
 
     @Override
-    public void deleteProduct(Long id) {
+    public void deleteProduct(Long id) throws ProductNotFoundException {
+        log.info("Deleting product with id:{}",id);
+
+        Optional<Product> product = productRepository.findById(id);
+
+        if(!product.isPresent()){
+            throw new ProductNotFoundException("Product not available");
+        }
         productRepository.deleteById(id);
     }
 }
